Improved UX by showing all hex values by default
I'm working to explore a few different color scales for our design system at AngelList and this tool has been a huge time saver. Unfortunately, the feature of hiding the Hex codes was frustrating, because I wanted to take screenshots of the scale that include all the hex codes. I used a CSS injector extension to fix this for myself, but I also wanted to provide a PR to propose this "fix" to the creator. 🤞